+++ src/Casts/DateCast.php
@@ -2,10 +2,9 @@
 
 namespace Naykel\Gotime\Casts;
 
-use Carbon\Carbon;
-use Illuminate\Support\Facades\Validator;
-
 use Illuminate\Contracts\Database\Eloquent\CastsAttributes;
+use Illuminate\Database\Eloquent\Model;
+use Carbon\Carbon;
 
 class DateCast implements CastsAttributes
 {
@@ -14,18 +13,22 @@ class DateCast implements CastsAttributes
      * Cast the dates to human readable dates
      * IMPORTANT must be a 'datetime' or 'timestamp'
      */
-    public function get($model, $key, $value, $attributes)
+    public function get(Model $model, string $key, mixed $value, array $attributes): string
     {
-
-        return $value ? Carbon::createFromFormat('Y-m-d H:i:s', $value)->format('d-m-Y') : null;
+        // 16-02-2022
+        return $value ? Carbon::createFromFormat('Y-m-d H:i:s', $value)->format('d-m-y') : null;
+        // Feb 16, 2022
+        // return $value ? Carbon::createFromFormat('Y-m-d H:i:s', $value)->format('M d, Y') : null;
     }
 
     /**
      * Converts to data back to database format
      */
-    public function set($model, $key, $value, $attributes)
+    public function set(Model $model, string $key, mixed $value, array $attributes): Carbon
     {
-
         return $value ? Carbon::parse($value) : null;
     }
 }
+
+
+